fix(artifact): add missing description field to Artifact types
Connect artifact-service persistence in create_artifact handler

Deduplicate shared constants and utility functions to @proma/shared
